+++ DataStructureVisualizer/include/Model/Frame.h
@@ -7,15 +7,15 @@
 class Frame {
     private:
     std::vector<int> currentData;
-    std::vector<int> highlightIndices;
+    std::vector<size_t> highlightIndices;
     int codeLineId;
     std::string message;
 
     public:
-    Frame(const std::vector<int>& data, const std::vector<int>& highlight, int lineId, const std::string& msg);
+    Frame(const std::vector<int>& data, const std::vector<size_t>& highlight, int lineId, const std::string& msg);
 
     const std::vector<int>& getCurrentData() const;
-    const std::vector<int>& getHighlightIndices() const;
+    const std::vector<size_t>& getHighlightIndices() const;
     int getCodeLineId() const;
     const std::string& getMessage() const;
 };
